requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
fastapi==0.128.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mcp==1.25.0
nest-asyncio==1.6.0
packaging==25.0
playwright==1.55.0
pluggy==1.6.0
proto-plus==1.27.0
protobuf==5.29.5
pyaes==1.6.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-multipart==0.0.21
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
SQLAlchemy==2.0.45
sqlmodel==0.0.31
sse-starlette==3.2.0
starlette==0.50.0
Telethon==1.42.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
